requirements.txt
Flask
requests
python-dotenv
psycopg2-binary
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
